+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version 1.1.1
+
++ Improved `mec_blocking()` by using inverted unsupervised MEC.
++ Added `alpha` in `mec_blocking()` for controlling the FLR-MMR trade-off. 
+
 # Version 1.1.0
 
 + Added `mec_blocking()` for blocked unsupervised MEC with pooled training